.internal/deployer/pyproject.toml
[project]
name = "deployer"
version = "0.1.0"
description = ""
authors = [
    { name = "Andrew Moffat", email = "arwmoffat@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[project.scripts]
deployer = "deployer:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
